requirements.txt
aiosmtpd==1.4.6
alembic==1.14.1
atpublic==5.1
attrs==25.1.0
bcrypt==4.2.1
blinker==1.9.0
click==8.1.8
collectionpy==1.2.0
colorama==0.4.6
dnspython==2.7.0
email_validator==2.2.0
Flask==3.1.0
Flask-Bcrypt==1.0.1
Flask-Login==0.6.3
Flask-Mail==0.10.0
Flask-Migrate==4.1.0
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.2
greenlet==3.1.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.5
Mako==1.3.9
MarkupSafe==3.0.2
numpy==2.2.3
PyJWT==2.1.0
python-dotenv==1.0.1
SQLAlchemy==2.0.38
typing_extensions==4.12.2
Werkzeug==3.1.3
wtf==0.1
WTForms==3.2.1
